pyproject.toml
[project]
name = "data-pipeline-agent"
version = "0.1.0"
description = "Planner-Executor Pattern demo"
authors = [
    {name = "AntonioDaria"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "anthropic (>=0.89.0,<0.90.0)",
    "pandas (>=3.0.2,<4.0.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
